+++ src/modules/auth/auth.controller.ts
@@ -1,10 +1,16 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, HttpCode, HttpStatus, HttpException } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, HttpCode, HttpStatus, HttpException, Res } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthLoginDto } from './dto/auth.dto';
+import { CreateUserDto } from '../users/dto/create-user.dto';
 
 @Controller('auth')
 export class AuthController {
   constructor(private readonly authService: AuthService) {}
+
+  @Post('register')
+  create(@Res({ passthrough: true}) res: Response, @Body() createUserDto: CreateUserDto) {
+    return this.authService.signUp(createUserDto);
+  }
 
   @Post('login')
   @HttpCode(HttpStatus.OK)
